add tree styles
