--- src/components/pages/tree/tree.scss
@@ -1,38 +1,0 @@
-.app-tree {
-  div {
-
-    div {
-      margin-left: 10px;
-      position: relative;
-
-      &::before {
-        content: ' ';
-        display: block;
-        position: absolute;
-        margin-left: -5px;
-        background-color: #ef0000;
-        width: 5px;
-        height: 2px;
-        top: 0.5em;
-      }
-
-      &::after {
-        content: ' ';
-        display: block;
-        position: absolute;
-        margin-left: -5px;
-        background-color: #ef0000;
-        width: 2px;
-        height: calc(100% + 0.5em);
-        top: 0;
-      }
-
-      &:last-child {
-        &::after {
-          width: 2px;
-          height: 0.5em;
-        }
-      }
-    }
-  }
-}
