+++ src/pages/family-table-page/family-table-page.scss
