+++ src/pages/main-page/index.js
